Added logic for special characters

+++ src/main/java/io/fi0x/languagegenerator/db/entities/Language.java
@@ -22,5 +22,5 @@ public class Language
     private int charsAfterSpecial;
     private int minSpecialChars;
     private int maxSpecialChars;
-    private Float specialCharacterChance;
+    private Double specialCharacterChance;
 }
